+++ src/js/packages/@reactpy/client/src/reactpy-client.ts
@@ -204,6 +204,39 @@ function createReconnectingWebSocket(
     socket.current.onopen = () => {
       everConnected = true;
       logger.log("client connected");
+
+      let _localStorage : any = {}
+      for (let i = 0; i <= window.localStorage.length; i++) {
+        let storage_key = window.localStorage.key(i)
+        if (storage_key){
+          _localStorage[storage_key] = window.localStorage.getItem(storage_key)
+        }
+      }
+      socket.current?.send(
+        JSON.stringify(
+          {
+            "type": "sync-local-storage",
+            "storage": _localStorage
+          }
+        )
+      )
+
+      let _sessionStorage : any = {}
+      for (let i = 0; i <= window.sessionStorage.length; i++) {
+        let storage_key = window.sessionStorage.key(i)
+        if (storage_key){
+          _sessionStorage[storage_key] = window.sessionStorage.getItem(storage_key)
+        }
+      }
+      socket.current?.send(
+        JSON.stringify(
+          {
+            "type": "sync-session-storage",
+            "storage": _sessionStorage
+          }
+        )
+      )
+
       interval = startInterval;
       retries = 0;
       if (props.onOpen) {