+++ src/py/reactpy/reactpy/backend/hooks.py
@@ -5,7 +5,7 @@ from typing import Any
 
 from reactpy.backend.types import Connection, Location
 from reactpy.core.hooks import create_context, use_context
-from reactpy.core.types import Context
+from reactpy.core.types import Context, LocalStorage, SessionStorage
 
 # backend implementations should establish this context at the root of an app
 ConnectionContext: Context[Connection[Any] | None] = create_context(None)
@@ -28,3 +28,11 @@ def use_scope() -> MutableMapping[str, Any]:
 def use_location() -> Location:
     """Get the current :class:`~reactpy.backend.types.Connection`'s location."""
     return use_connection().location
+
+def use_local_storage() -> LocalStorage:
+    """Get the localStorage object for the connection"""
+    return use_connection().local_storage
+
+def use_session_storage() -> SessionStorage:
+    """Get the sessionStorage object for the connection"""
+    return use_connection().session_storage