+++ src/py/reactpy/reactpy/core/serve.py
@@ -9,8 +9,16 @@ from anyio import create_task_group
 from anyio.abc import TaskGroup
 
 from reactpy.config import REACTPY_DEBUG_MODE
-from reactpy.core.types import LayoutEventMessage, LayoutType, LayoutUpdateMessage
-
+from reactpy.core.types import (
+    LocalStorageEventMessage, 
+    SessionStorageEventMessage, 
+    LocalStorage, 
+    SessionStorage,
+    LayoutEventMessage, 
+    LayoutType, 
+    LayoutUpdateMessage
+)
+from reactpy.core.layout import Layout
 logger = getLogger(__name__)
 
 
@@ -36,6 +44,8 @@ class Stop(BaseException):
 
 async def serve_layout(
     layout: LayoutType[LayoutUpdateMessage, LayoutEventMessage],
+    local_storage: LocalStorage,
+    session_storage: SessionStorage,
     send: SendCoroutine,
     recv: RecvCoroutine,
 ) -> None:
@@ -44,15 +54,9 @@ async def serve_layout(
         try:
             async with create_task_group() as task_group:
                 task_group.start_soon(_single_outgoing_loop, layout, send)
-                task_group.start_soon(_single_incoming_loop, task_group, layout, recv)
-        except Stop:  # nocov
-            warn(
-                "The Stop exception is deprecated and will be removed in a future version",
-                UserWarning,
-                stacklevel=1,
-            )
+                task_group.start_soon(_single_incoming_loop, task_group, layout, local_storage, session_storage,recv)
+        except Stop:
             logger.info(f"Stopped serving {layout}")
-
 
 async def _single_outgoing_loop(
     layout: LayoutType[LayoutUpdateMessage, LayoutEventMessage], send: SendCoroutine
@@ -71,13 +75,27 @@ async def _single_outgoing_loop(
                 logger.error(msg)
             raise
 
+async def incoming_router(
+    layout: Layout,
+    local_storage: LocalStorage,
+    session_storage: SessionStorage,
+    event: LayoutEventMessage or LocalStorageEventMessage or SessionStorageEventMessage,
+): 
+    if event["type"] == "sync-local-storage":
+        local_storage._sync(event["storage"])
+    elif event["type"] == "sync-session-storage":
+        session_storage._sync(event["storage"])
+    elif event["type"] == "layout-event":
+        await layout.deliver(event)
 
 async def _single_incoming_loop(
     task_group: TaskGroup,
     layout: LayoutType[LayoutUpdateMessage, LayoutEventMessage],
+    local_storage: LocalStorage,
+    session_storage: SessionStorage,
     recv: RecvCoroutine,
 ) -> None:
     while True:
         # We need to fire and forget here so that we avoid waiting on the completion
         # of this event handler before receiving and running the next one.
-        task_group.start_soon(layout.deliver, await recv())
+        task_group.start_soon(incoming_router, layout, local_storage, session_storage, await recv())