+++ docs/source/about/changelog.rst
@@ -50,6 +50,8 @@ Unreleased
 - :pull:`1171` - The ``Stop`` exception. Recent releases of ``anyio`` have made this
   exception difficult to use since it now raises an ``ExceptionGroup``. This exception
   was primarily used for internal testing purposes and so is now deprecated.
+**Added**
+- :issue:`1075` - Add support to localStorage and sessionStorage access
 
 
 v1.0.2